--- build_external/scenarios/gaps_hi/child-compose.yml
@@ -1,14 +1,0 @@
-version: '3.3'
-services:
-  agent_child:
-    image: debian_10_dev
-    command: /usr/sbin/netdata -D -W set global "debug flags" 0xc0000000 -W set global hostname "child"
-    ports:
-    - 21002:19999
-    volumes:
-      - ./child_stream.conf:/etc/netdata/stream.conf:ro
-      #- ./child_guid:/var/lib/netdata/registry/netdata.public.unique.id:ro
-      - ./min.conf:/etc/netdata/netdata.conf:ro
-    cap_add:
-    - SYS_PTRACE
-
