--- build_external/scenarios/gaps_hi/child_guid
@@ -1,1 +1,0 @@
-22222222-2222-2222-2222-222222222222
