--- build_external/scenarios/gaps_hi/child_stream.conf
@@ -1,10 +1,0 @@
-[stream]
-    enabled = yes
-    destination = tcp:agent_middle
-    api key = 00000000-0000-0000-0000-000000000000
-    timeout seconds = 60
-    default port = 19999
-    send charts matching = *
-    buffer size bytes = 10485760
-    reconnect delay seconds = 5
-    initial clock resync iterations = 60
