--- build_external/scenarios/gaps_hi/master-cert.pem
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDqzCCApOgAwIBAgIUfvGLZNcgRcPxvZUIE7x+14egqPMwDQYJKoZIhvcNAQEN
-BQAwZTELMAkGA1UEBhMCU0UxEDAOBgNVBAgMB1RoZU1vb24xEDAOBgNVBAcMB0x1
-bmFPbmUxDzANBgNVBAoMBk5ldGF0YTEQMA4GA1UECwwHTmV0ZGF0YTEPMA0GA1UE
-AwwGbWFzdGVyMB4XDTIwMDYwMjEwMzM0OVoXDTIxMDYwMjEwMzM0OVowZTELMAkG
-A1UEBhMCU0UxEDAOBgNVBAgMB1RoZU1vb24xEDAOBgNVBAcMB0x1bmFPbmUxDzAN
-BgNVBAoMBk5ldGF0YTEQMA4GA1UECwwHTmV0ZGF0YTEPMA0GA1UEAwwGbWFzdGVy
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEApJugxJOyN03fNFzduY1m
-PNx9V0g6ljM+8DEmGiUVG0m/njGBClawJgvaO+xE9JAZDT61HwGfBUK4ad4w38Of
-1BN0curleKLdu/y/xi+GKZXBWlGgelAssrtMEWDqOdIdkbndZCfpWhsc2ODvWUiR
-U83XnuZ99g/21U2SrRdYobb20fYauMWdNK8R2rbqBIJh8yTAKwLymufwoQVybiLs
-uhUIiH5voyFfplbh2kbLYazQoFbf+eamjQoU+TeCrsP/o1W7yn+C1RcBjOW8KFTW
-T4mGgMSvBFJiIW8lSL/ezH63eOlI+bYXebRNhebaF3hp9PYNPcWlwWR+fC8MdtGL
-JQIDAQABo1MwUTAdBgNVHQ4EFgQUM7f4DWfSDb0WWDCv6cYjobrU8n0wHwYDVR0j
-BBgwFoAUM7f4DWfSDb0WWDCv6cYjobrU8n0wDwYDVR0TAQH/BAUwAwEB/zANBgkq
-hkiG9w0BAQ0FAAOCAQEAEh2C56xAy6DJGBFCtuunwD1aUsgiSseebRt+W3bonX05
-TH7CL7lxIyg0j3FodDMI3nT4FXe85xuFZ669wJZ8iGtSgXdRFMFSgcF2kAcF+xZn
-eJI8JZkVVGtrJG5l8ir8ooUWM5FjGAqQz5K+UM7xtYnqVHw8UvJPS5YZgJx+Nqis
-I7Yc8pGynXwZ5bjO4Uoveyghe2HG+plinwG7iyQxY+V8F2uW6kP4v/PjdMzzjhC2
-vfhXmLvHVukLnGYYhtEDA1lvqniy+4+pXwpDDXmotwgkib5g8+q+zKuXoFPekRuv
-EjnPxbT+zj4NYN2gv06U1wWPuTlwSXGztORdN9MZEA==
------END CERTIFICATE-----
