--- build_external/scenarios/gaps_hi/master-key.pem
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCkm6DEk7I3Td80
-XN25jWY83H1XSDqWMz7wMSYaJRUbSb+eMYEKVrAmC9o77ET0kBkNPrUfAZ8FQrhp
-3jDfw5/UE3Ry6uV4ot27/L/GL4YplcFaUaB6UCyyu0wRYOo50h2Rud1kJ+laGxzY
-4O9ZSJFTzdee5n32D/bVTZKtF1ihtvbR9hq4xZ00rxHatuoEgmHzJMArAvKa5/Ch
-BXJuIuy6FQiIfm+jIV+mVuHaRsthrNCgVt/55qaNChT5N4Kuw/+jVbvKf4LVFwGM
-5bwoVNZPiYaAxK8EUmIhbyVIv97Mfrd46Uj5thd5tE2F5toXeGn09g09xaXBZH58
-Lwx20YslAgMBAAECggEALqyuVm8mzoW3L6rReCa/AaznoZMqgaKASy00h59SL82m
-lt2koGDaS796nxksqEgfwmCwjN0x0RUNI3/MTfazYDgwAdHXRXCw/Pf7DOA2453C
-GLEiluwa8IcpV5gqckDA/6w7OYYTpeggMuL9G0z3YFLyu+zRdK8sJ2l6mLgQ+rFp
-G/6Q+BkyNBSfTsMNW5nWxxfhcGiobMvc6JCvNXGl0VpVqJE4MUkiaPPgTDCzs3+H
-vqrOIaUrnTBEJXrXdCfOlsXZ+amGHvus2BPELt1+6/gKk5+kNWTRLczxBuskaR/p
-FVPGAc2BG+O/e8sJL6vVtE9uuJJ6H1p9LTT1n6NWlQKBgQDW1QdQBnR0BJBhb3ZT
-p8V1zGPqOhiexnUk0+picDr1BjZ2m4XjX1a2EYnOhXusohEO9S1L1d+kTlGil7By
-7YjpWodcJX70jjlq0FWnRRh0AswHyhR1Z4ugNc1Vq1qHWPiTkXxIBvFmDB4MQxyZ
-+SyXhZyAIoojhKRXXTSOyI/HzwKBgQDEJsLjWLOguAJRpfl804anSTKzeC5M+hgn
-ZGLMIRsZwcf2oTtg9bRggxE/6NguaLhFAa6d2Ly5jsZQjO+fXrb+AMdrkUgjlpcc
-7qvu4ngF+uzETQz+wr4nQx86+S7ocLkYTSLeyiLBzWa1bT2Y80Jy+pUQL2zmZ7kw
-SUhlkHzGywKBgQCVVyVsNfCCLiIH+anx4XDcwKa8ToKK7T7FySLoJ0h0yErYDKj4
-NHBcxLhdfmFI0qi1Ey7XxudemqsZMgWQ/xrrzOjH37DvZGb5n7XLP7cN/lvDFnE+
-ZCqlxoDlCuc1q12xh7MzhtAMtmO09npNdu/OrSzrzhblNfHgYYSFtVpt9QKBgBeu
-+m8K4b1xGk6TdLk+rQ57xxRmWWjP9aSj+U3eXXBRtnwv4YSGhj309Vs7GQcZe85R
-Mq06Yycgdw94xOIARKYMoAd75jLt1tcWdkoErCtErdvR1dTY3K7O6vFq7t//5Elr
-3t75lxLBYzpas0pJbOeHHAJ1R3fWZETRRF2HAZnbAoGAOY4RmxAlv3QUZsQvIbhK
-o9SagjvwsO14SDY/qmBrbCkmIEGPGTuo8qQzJGx9G+Sh2tZSMjNx3KZvJgTpw+d7
-ObGaR7P6uTQZm4E9S6OhUTYeJkCHZsVbp3T+vecew1m+DXTS4CLB3BYzJz3FPEoF
-RiSUGeRDdViCIaBC1C1yd+I=
------END PRIVATE KEY-----
