--- build_external/scenarios/gaps_hi/middle-compose.yml
@@ -1,13 +1,0 @@
-version: '3.3'
-services:
-  agent_middle:
-    image: debian_10_dev
-    command: /usr/sbin/netdata -D -W set global "debug flags" 0xc0020000 -W set global hostname "middle"
-    ports:
-      - 21001:19999
-    volumes:
-      - ./middle_stream.conf:/etc/netdata/stream.conf:ro
-      - ./middle_guid:/var/lib/netdata/registry/netdata.public.unique.id:ro
-      - ./min.conf:/etc/netdata/netdata.conf:ro
-    cap_add:
-    - SYS_PTRACE
