--- build_external/scenarios/gaps_hi/middle_guid
@@ -1,1 +1,0 @@
-11111111-1111-1111-1111-111111111111
