--- build_external/scenarios/gaps_hi/middle_stream.conf
@@ -1,23 +1,0 @@
-[stream]
-    enabled = yes
-    destination = tcp:agent_parent
-    api key = 00000000-0000-0000-0000-000000000000
-    timeout seconds = 60
-    default port = 19999
-
-    send charts matching = *
-    buffer size bytes = 1048576
-    reconnect delay seconds = 5
-    initial clock resync iterations = 60
-
-[00000000-0000-0000-0000-000000000000]
-    enabled = yes
-    allow from = *
-    default history = 3600
-    # default memory mode = ram
-
-    health enabled by default = auto
-
-    # postpone alarms for a short period after the sender is connected
-    default postpone alarms on connect seconds = 60
-    multiple connections = allow
