--- build_external/scenarios/gaps_hi/min.conf
@@ -1,7 +1,0 @@
-[global]
-    debug flags = 0x00000000c0000000
-    errors flood protection period = 0
-    hostname = wintermute
-[web]
-    ssl key = /etc/netdata/ssl/key.pem
-    ssl certificate = /etc/netdata/ssl/cert.pem
