--- build_external/scenarios/gaps_hi/parent-compose.yml
@@ -1,13 +1,0 @@
-version: '3.3'
-services:
-  agent_parent:
-    image: debian_10_dev
-    command: /usr/sbin/netdata -D -W set global "debug flags" 0xc0000000 -W set global hostname "parent"
-    ports:
-      - 21000:19999
-    volumes:
-      - ./parent_stream.conf:/etc/netdata/stream.conf:ro
-      - ./parent_guid:/var/lib/netdata/registry/netdata.public.unique.id:ro
-      - ./min.conf:/etc/netdata/netdata.conf:ro
-    cap_add:
-      - SYS_PTRACE
