--- build_external/scenarios/gaps_hi/parent_guid
@@ -1,1 +1,0 @@
-00000000-0000-0000-0000-000000000000
