--- build_external/scenarios/gaps_hi/parent_stream.conf
@@ -1,11 +1,0 @@
-[00000000-0000-0000-0000-000000000000]
-    enabled = yes
-    allow from = *
-    default history = 3600
-    default memory mode = dbengine
-    health enabled by default = no
-
-    # postpone alarms for a short period after the sender is connected
-    default postpone alarms on connect seconds = 60
-    multiple connections = allow
-
